Add yaml tree-sitter as a submodule
Tested by following the following instructions:
https://github.com/ubolonton/tree-sitter-langs#building-grammars-from-source
as well as by using this as a submodule of emacs-tree-sitter.

The grammar compiles and loads correctly. Queries can be built using
`tree-sitter-query-builder`.

This commit does not add a `highlights.scm` file, this work is pending
for yaml support. Relates to / Implements #21

Repo: https://github.com/ikatyang/tree-sitter-yaml.git
Commit:
https://github.com/ikatyang/tree-sitter-yaml/commit/0e36bed171768908f331ff7dff9d956bae016efb